Extract MarkdownRenderer component
